1、修改copyBean时PageCondition对象pageNo，pageSize拷贝失败的问题。

See merge request architecture/kcwl-ddd-framework!88

+++ kcwl-ddd-core/src/main/java/com/kcwl/ddd/domain/entity/PageCondition.java
@@ -11,7 +11,7 @@ public class PageCondition extends Criteria {
         return curPagerNo;
     }
 
-    public void setCurPagerNo(int curPagerNo) {
+    public void setCurPagerNo(long curPagerNo) {
         this.curPagerNo = curPagerNo;
     }
 
@@ -19,7 +19,7 @@ public class PageCondition extends Criteria {
         return pageSize;
     }
 
-    public void setPageSize(int pageSize) {
+    public void setPageSize(long pageSize) {
         this.pageSize = pageSize;
     }
